增加int类型，增加var和global关键字
增加int(64位)类型
增加var关键字，等效于local
增加global关键字用于声明全局变量
忽略浮点数结尾的'f‘,如"1.0f"的'f'会被正常解析并忽略.
JetContext增加OutputFunction成员用于信息输出

+++ Libraries/File.cpp
@@ -36,8 +36,8 @@ void RegisterFileLibrary(JetContext* context)
 			throw RuntimeException("Tried to read from a closed file!");
 
 		char* out = new char[((int)v->value)+1];//context->GCAllocate((v)->value);
-		fread(out, 1, (int)v[1], v[0].GetUserdata<FILE>());
-		out[(int)v[1]] = 0;
+		fread(out, 1, (int64_t)v[1], v[0].GetUserdata<FILE>());
+		out[(int64_t)v[1]] = 0;
 		return context->NewString(out, false);
 	});
 	(*filePrototype)["Write"] = Value([](JetContext* context, Value* v, int args)
